krejoinstack/requirements.txt
paramiko
oslo_concurrency
pyyaml
